requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
prompt-toolkit==3.0.43
python-vlc==3.0.20123
requests==2.31.0
urllib3==2.2.1
wcwidth==0.2.13
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
prompt-toolkit==3.0.43
pyfiglet==1.0.2
python-vlc==3.0.20123
requests==2.31.0
urllib3==2.2.1
wcwidth==0.2.13
